Update and Rename Mongo.js to Firestore.js
